fix: 포인트 테스트 파일 분리
