+++ apps/commerce-api/src/test/java/com/loopers/interfaces/api/user/UserV1ApiE2ETest.java
@@ -42,8 +42,6 @@ public class UserV1ApiE2ETest {
     private static final String EMAIL = "donghee@test.com";
     private static final String SIGN_UP_ENDPOINT = "/api/v1/users";
     private static final String INFO_CHECK_ENDPOINT = "/api/v1/users/me";
-    private static final String POINT_CHECK_ENDPOINT = "/api/v1/points";
-    private static final String POINT_CHARGE_ENDPOINT = "/api/v1/points/charge";
 
     @AfterEach
     void tearDown() {
@@ -140,101 +138,6 @@ public class UserV1ApiE2ETest {
             ParameterizedTypeReference<ApiResponse<UserV1Dto.UserResponse>> responseType = new ParameterizedTypeReference<>() {};
             ResponseEntity<ApiResponse<UserV1Dto.UserResponse>> response =
                     testRestTemplate.exchange(INFO_CHECK_ENDPOINT, HttpMethod.GET, httpEntity, responseType);
-
-            // assert
-            assertAll(
-                    () -> assertTrue(response.getStatusCode().is4xxClientError()),
-                    () -> assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND)
-            );
-        }
-
-    }
-
-    @DisplayName("GET /api/v1/points")
-    @Nested
-    class PointCheck {
-        @DisplayName("포인트 조회에 성공할 경우, 보유 포인트를 응답으로 반환한다.")
-        @Test
-        void returnsPoints_whenPointCheckIsSuccessful() {
-            // arrange
-            testRestTemplate.postForEntity(SIGN_UP_ENDPOINT, defaultUserRequest(), Void.class);
-
-            HttpHeaders headers = new HttpHeaders();
-            headers.add("X-USER-ID", USER_ID);
-            HttpEntity<Void> httpEntity = new HttpEntity<>(headers);
-
-            // act
-            ParameterizedTypeReference<ApiResponse<Long>> responseType = new ParameterizedTypeReference<>() {
-            };
-            ResponseEntity<ApiResponse<Long>> response =
-                    testRestTemplate.exchange(POINT_CHECK_ENDPOINT, HttpMethod.GET, httpEntity, responseType);
-
-            // assert
-            assertAll(
-                    () -> assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK),
-                    () -> assertThat(response.getBody().data()).isEqualTo(0L)
-            );
-        }
-
-        @DisplayName("X-USER-ID 헤더가 없을 경우, 400 Bad Request 응답을 반환한다.")
-        @Test
-        void returns404BadRequest_whenXUSERIDDoesNotExist() {
-            // arrange
-            HttpEntity<Void> httpEntity = new HttpEntity<>(null);
-
-            // act
-            ParameterizedTypeReference<ApiResponse<Long>> responseType = new ParameterizedTypeReference<>() {
-            };
-            ResponseEntity<ApiResponse<Long>> response =
-                    testRestTemplate.exchange(POINT_CHECK_ENDPOINT, HttpMethod.GET, httpEntity, responseType);
-
-            // assert
-            assertAll(
-                    () -> assertTrue(response.getStatusCode().is4xxClientError()),
-                    () -> assertThat(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST)
-            );
-        }
-    }
-
-    @DisplayName("POST /api/v1/points/charge")
-    @Nested
-    class PointCharge {
-        @DisplayName("존재하는 유저가 1000원을 충전할 경우, 충전된 보유 총량을 응답으로 반환한다.")
-        @Test
-        void returnsTotalPoints_whenUserAddsPoints1000() {
-            // arrange
-            testRestTemplate.postForEntity(SIGN_UP_ENDPOINT, defaultUserRequest(), Void.class);
-
-            HttpHeaders headers = new HttpHeaders();
-            headers.add("X-USER-ID", USER_ID);
-            PointV1Dto.PointChargeRequest request = new PointV1Dto.PointChargeRequest(1000L);
-            HttpEntity<PointV1Dto.PointChargeRequest> httpEntity = new HttpEntity<>(request, headers);
-
-            // act
-            ParameterizedTypeReference<ApiResponse<Long>> responseType = new ParameterizedTypeReference<>() {};
-            ResponseEntity<ApiResponse<Long>> response =
-                    testRestTemplate.exchange(POINT_CHARGE_ENDPOINT, HttpMethod.POST, httpEntity, responseType);
-
-            // assert
-            assertAll(
-                    () -> assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK),
-                    () -> assertThat(response.getBody().data()).isEqualTo(1000L)
-            );
-        }
-
-        @DisplayName("존재하지 않는 유저로 요청할 경우, 404 Not Found 응답을 반환한다.")
-        @Test
-        void returns404NotFound_whenUserDoesNotExist() {
-            // arrange
-            HttpHeaders headers = new HttpHeaders();
-            headers.add("X-USER-ID", USER_ID);
-            PointV1Dto.PointChargeRequest request = new PointV1Dto.PointChargeRequest(1000L);
-            HttpEntity<PointV1Dto.PointChargeRequest> httpEntity = new HttpEntity<>(request, headers);
-
-            // act
-            ParameterizedTypeReference<ApiResponse<Long>> responseType = new ParameterizedTypeReference<>() {};
-            ResponseEntity<ApiResponse<Long>> response =
-                    testRestTemplate.exchange(POINT_CHARGE_ENDPOINT, HttpMethod.POST, httpEntity, responseType);
 
             // assert
             assertAll(
